Import Goal class and check linters
